apps/brain/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
openai
langchain
langchain-openai
pydantic
httpx
azure-identity
msgraph-core
msgraph-sdk
pydantic-settings
cryptography
# Document Processing
pdfplumber==0.10.3
openpyxl==3.1.2
pandas
python-multipart==0.0.6
